Adicionando Exercicio Casa

+++ ListaDeExercicios4Pt1-OrientacaoAObjetos/Exercicio7/ModeloDeComputador.java
@@ -55,7 +55,7 @@ public class ModeloDeComputador {
         return monitor;
     }
 	
-    public double calcularPreco(){
+    public int calcularPreco(){
         return placaMae + processador + memoria + discoRigido + monitor;
     }
 }
